fix : calendar fragment functions
