+++ app/src/main/java/coms/example/fivesensediary/decorators/EventDecorator.java
@@ -3,6 +3,8 @@ package coms.example.fivesensediary.decorators;
 import android.app.Activity;
 import android.graphics.drawable.Drawable;
 
+import androidx.fragment.app.Fragment;
+
 import coms.example.fivesensediary.R;
 import com.prolificinteractive.materialcalendarview.CalendarDay;
 import com.prolificinteractive.materialcalendarview.DayViewDecorator;
@@ -16,7 +18,7 @@ public class EventDecorator implements DayViewDecorator {
     private int color;
     private HashSet<CalendarDay> dates;
 
-    public EventDecorator(int color, Collection<CalendarDay> dates,Activity context) {
+    public EventDecorator(int color, Collection<CalendarDay> dates, Activity context) {
         drawable = context.getResources().getDrawable(R.drawable.more);
         this.color = color;
         this.dates = new HashSet<>(dates);
