+++ app/src/main/java/coms/example/fivesensediary/ui/sharedDiary/SharedDiaryFragment.java
@@ -1,12 +1,13 @@
 package coms.example.fivesensediary.ui.sharedDiary;
 
 import android.database.Cursor;
+import android.graphics.Color;
+import android.os.AsyncTask;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.TextView;
 import android.widget.Toast;
 
 import androidx.annotation.NonNull;
@@ -20,18 +21,19 @@ import com.prolificinteractive.materialcalendarview.CalendarMode;
 import com.prolificinteractive.materialcalendarview.MaterialCalendarView;
 import com.prolificinteractive.materialcalendarview.OnDateSelectedListener;
 
+import java.util.ArrayList;
 import java.util.Calendar;
+import java.util.List;
 import java.util.concurrent.Executors;
 
-import coms.example.fivesensediary.MainActivity;
 import coms.example.fivesensediary.R;
+import coms.example.fivesensediary.decorators.EventDecorator;
 import coms.example.fivesensediary.decorators.OneDayDecorator;
 import coms.example.fivesensediary.decorators.SaturdayDecorator;
 import coms.example.fivesensediary.decorators.SundayDecorator;
 
 public class SharedDiaryFragment extends Fragment {
 
-    String time,kcal,menu;
     private final OneDayDecorator oneDayDecorator = new OneDayDecorator();
     Cursor cursor;
     MaterialCalendarView materialCalendarView;
@@ -60,6 +62,8 @@ public class SharedDiaryFragment extends Fragment {
 
         String[] result = {"2021,01,11","2021,02,11","2021,03,11","2021,04,11"};
 
+        new ApiSimulator(result).executeOnExecutor(Executors.newSingleThreadExecutor());
+
         materialCalendarView.setOnDateChangedListener(new OnDateSelectedListener() {
             @Override
             public void onDateSelected(@NonNull MaterialCalendarView widget, @NonNull CalendarDay date, boolean selected) {
@@ -87,5 +91,51 @@ public class SharedDiaryFragment extends Fragment {
             }
         });
         return root;
+    }
+
+    private class ApiSimulator extends AsyncTask<Void, Void, List<CalendarDay>> {
+
+        String[] Time_Result;
+
+        ApiSimulator(String[] Time_Result){
+            this.Time_Result = Time_Result;
+        }
+
+        @Override
+        protected List<CalendarDay> doInBackground(@NonNull Void... voids) {
+            try {
+                Thread.sleep(500);
+            } catch (InterruptedException e) {
+                e.printStackTrace();
+            }
+
+            Calendar calendar = Calendar.getInstance();
+            ArrayList<CalendarDay> dates = new ArrayList<>();
+
+            /*특정날짜 달력에 점표시해주는곳*/
+            /*월은 0이 1월 년,일은 그대로*/
+            //string 문자열인 Time_Result 을 받아와서 ,를 기준으로짜르고 string을 int 로 변환
+            for(int i = 0 ; i < Time_Result.length ; i ++){
+                CalendarDay day = CalendarDay.from(calendar);
+                String[] time = Time_Result[i].split(",");
+                int year = Integer.parseInt(time[0]);
+                int month = Integer.parseInt(time[1]);
+                int dayy = Integer.parseInt(time[2]);
+
+                dates.add(day);
+                calendar.set(year,month-1,dayy);
+            }
+
+
+
+            return dates;
+        }
+
+        @Override
+        protected void onPostExecute(@NonNull List<CalendarDay> calendarDays) {
+            super.onPostExecute(calendarDays);
+
+            materialCalendarView.addDecorator(new EventDecorator(Color.GREEN, calendarDays,getActivity()));
+        }
     }
 }
